requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-auth==2.16.0
google-auth-oauthlib==1.0.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.14
wrapt==1.14.1
